% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchShinyMultiSDA}
\alias{launchShinyMultiSDA}
\title{Launch ShinySDA.}
\usage{
launchShinyMultiSDA(...)
}
\value{
Shiny application.
}
\description{
Launch ShinySDA.
}
\details{
Launch ShinyMultiSDA.
}
\keyword{SDA}
\keyword{shiny}
